persistence test in progress
