+++ src/main/webapp/js/pages/tests/casperjs/README.md
@@ -31,7 +31,7 @@ To run Core default projects Tests (Requires NOT having the persistence bundle) 
 
 To run Persistence Tests (Requires the persistence bundle and a running MySQL server):
 
-`casperjs test PersistenceTests.js --engine=slimerjs` to run Geppetto Persistence Tests.
+`casperjs test --includes=PersistenceTestsLogic.js PersistenceTests.js --engine=slimerjs` to run Geppetto Persistence Tests.
 
 Tests are executed by default on port 8080. If you would like to execute tests on a different port you can change it [here](https://github.com/openworm/org.geppetto.frontend/blob/downloadData/src/main/webapp/js/pages/tests/casperjs/CoreTestsUtility.js#L1).
